+++ src/main/kotlin/Winter/pets/service/CenterService.kt
@@ -1,0 +1,6 @@
+package Winter.pets.service
+
+interface CenterService {
+    fun findToCenter(siName: String, gunguName: String): List<String>
+    fun addToCenter(siName: String, gunguName: String)
+}
