+++ src/main/kotlin/Winter/pets/service/GunguService.kt
@@ -1,0 +1,6 @@
+package Winter.pets.service
+
+interface GunguService {
+    fun findToGu(name: String): List<String>
+    fun addToGu(name: String)
+}
