+++ src/main/kotlin/Winter/pets/service/KindService.kt
@@ -1,0 +1,6 @@
+package Winter.pets.service
+
+interface KindService {
+    fun findToKind(kindName: String): List<String>
+    fun addToKind(kindName: String)
+}
