+++ src/main/kotlin/Winter/pets/service/SiService.kt
@@ -1,0 +1,5 @@
+package Winter.pets.service
+
+interface SiService {
+    fun addToCountry():Unit
+}
